Struct syntastica_parsers::LanguageSetImpl

pub struct LanguageSetImpl(_);

An implementation of [LanguageSet] including all languages in the enabled feature set.

Languages are loaded the first time they are requested and will then be reused for later accesses. To pre-load a list of languages, use preload or preload_all.

Implementations

impl LanguageSetImpl

pub fn new() -> Self

Create a new LanguageSetImpl with no pre-loaded languages.

pub fn preload(&mut self, languages: &[&str]) -> Result<()>

Pre-load the given list of languages.

To pre-load all supported languages, use preload_all.

Errors

If the languages list contains a name of a language that is not included in this set, an [Error::UnsupportedLanguage] error is returned and no further languages are loaded.

pub fn preload_all(&mut self)

Pre-load all languages in this set.

To pre-load a specific set of languages, use preload.