Crate syntastica_parsers

syntastica-parsers

Collection of tree-sitter parsers for syntastica, using parsers available on crates.io.

See the project overview for more information on all parser collections.

List of included parsers

List of parsers included in the some feature

bash (not supported by this collection)

c

cpp

css

go

html (not supported by this collection)

java

javascript

json

python

rust

tsx (not supported by this collection)

typescript (not supported by this collection)

List of parsers additionally included in the most feature

asm

regex (not supported by this collection)

List of parsers additionally included in the all feature

Features

At least one of some, most, and all must be enabled.

some — Include parsers for the most widely known supported languages.

most — Implies some. Include parsers for most common languages.

all — Implies most. Include parsers for all supported languages.

docs — Meant to be enabled when building docs

Structs

LanguageProviderImpl

An implementation of LanguageProvider, providing all parsers in the enabled feature set (see all) or a subset of them (see with_languages).

Functions

asmmost

Get the parser for asm.

bashsome

This parser is not supported by this parser collection and thus this function will panic!

csome

Get the parser for c.

cppsome

Get the parser for cpp.

csssome

Get the parser for css.

gosome

Get the parser for go.

htmlsome

This parser is not supported by this parser collection and thus this function will panic!

javasome

Get the parser for java.

javascriptsome

Get the parser for javascript.

jsonsome

Get the parser for json.

pythonsome

Get the parser for python.

regexmost

This parser is not supported by this parser collection and thus this function will panic!

rustsome

Get the parser for rust.

tsxsome

This parser is not supported by this parser collection and thus this function will panic!

typescriptsome

This parser is not supported by this parser collection and thus this function will panic!